Create Revel - Excersic11-17
(Algebra: perfect square)

Write a program that prompts the user to enter an integer m and find the smallest integer n such that m * n is a perfect square.

(Hint: Store all smallest factors of m into an array list. n is the product of the factors that appear an odd number of times in the array list.
For example, consider m = 90, store the factors 2, 3, 3, 5 in an array list. 2 and 5 appear an odd number of times in the array list. So, n is 10.)

Sample Run 1

Enter an integer m: 1500
The smallest number n for m x n to be a perfect square is 15
m x n is 22500